requirements.txt
certifi==2021.10.8
chardet==4.0.0
idna==2.10
PyYAML==5.4.1
requests==2.26.0
timeloop==1.0.2
urllib3==1.26.7
